+++ R/replace_mnirs.R
@@ -354,18 +354,12 @@ replace_invalid <- function(
             )
         }
 
-        window_idx <- compute_local_windows(
-            t,
-            invalid_idx,
-            width,
-            span,
-            env = env
-        )
-        local_medians <- compute_local_fun(y, window_idx, median_nona)
-        ## if method = "median"
+        bounds <- compute_window_bounds(t, invalid_idx, width, span, env = env)
         ## invalid_values removed to NA first,
         ## so returns local median excluding idx
-        y[invalid_idx] <- local_medians
+        y[invalid_idx] <- vapply(seq_along(invalid_idx), \(.i) {
+            median_no_na(y[bounds$start[.i]:bounds$end[.i]])
+        }, numeric(1))
     }
 
     return(y)
@@ -548,7 +542,7 @@ replace_missing <- function(
         y <- x
         na_idx <- which(is.na(x))
         window_idx <- compute_valid_neighbours(x, t, width, span, verbose, env)
-        local_medians <- compute_local_fun(x, window_idx, median_nona)
+        local_medians <- compute_local_fun(x, window_idx, median_no_na)
         y[na_idx] <- local_medians
     }