+++ R/validate_mnirs.R
@@ -576,7 +576,7 @@ validate_start_time <- function(
 
 #' wrap findInterval: informative 'time_channel' error message
 #' @keywords internal
-findInt_mnirs <- function(x, vec, ..., env = rlang::caller_env()) {
+validate_findInt <- function(x, vec, ..., env = rlang::caller_env()) {
     if (anyNA(vec) || is.unsorted(vec)) {
         cli_abort(c(
             "x" = "Irregular {.arg time_channel} samples detected.",